refactor: change folder structure
